metablock_iterator: use size_t instead of uint64_t for skip.

+++ include/sqsh_metablock_private.h
@@ -81,7 +81,7 @@ SQSH_NO_UNUSED int sqsh__metablock_iterator_init(
  * @return 0 on success, or a negative value on error.
  */
 SQSH_NO_UNUSED int sqsh__metablock_iterator_skip(
-		struct SqshMetablockIterator *iterator, uint64_t amount);
+		struct SqshMetablockIterator *iterator, size_t amount);
 
 /**
  * @internal
